+++ packages/vibrant-components/src/lib/ModalBottomSheet/ModalBottomSheet.spec.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, waitFor, waitForElementToBeRemoved } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { Box, PortalRootProvider } from '@vibrant-ui/core';
+import { Box, PortalRootProvider, VibrantProvider } from '@vibrant-ui/core';
 import type { ReactRenderer } from '@vibrant-ui/utils/testing-web';
 import { BrowserTesting, createReactRenderer } from '@vibrant-ui/utils/testing-web';
 import { Pressable } from '../Pressable';
@@ -244,7 +244,11 @@ describe('<ModalBottomSheet />', () => {
     beforeEach(async () => {
       await BrowserTesting.resize(639, 768);
 
-      renderer = render(<ModalBottomSheet open={true} renderContents={() => null} />);
+      renderer = render(
+        <VibrantProvider>
+          <ModalBottomSheet open={true} renderContents={() => null} />
+        </VibrantProvider>
+      );
     });
 
     it('match snapshot', async () => {
@@ -258,7 +262,11 @@ describe('<ModalBottomSheet />', () => {
     beforeEach(async () => {
       await BrowserTesting.resize(1024, 768);
 
-      renderer = render(<ModalBottomSheet open={true} renderContents={() => null} />);
+      renderer = render(
+        <VibrantProvider>
+          <ModalBottomSheet open={true} renderContents={() => null} />
+        </VibrantProvider>
+      );
     });
 
     it('match snapshot', async () => {