+++ packages/vibrant-components/src/lib/Radio/index.ts
@@ -1,0 +1,2 @@
+export { Radio } from './Radio';
+export { useRadio } from './useRadio';