+++ packages/vibrant-example-app/src/stories.ts
@@ -1,6 +1,7 @@
 export const stories = [
   require('../../vibrant-components-web/src/lib/CSSTransition/CSSTransition.stories.tsx'),
   require('../../vibrant-components/src/docs/Palettes.stories.tsx'),
+  require('../../vibrant-components/src/lib/Avatar/Avatar.stories.tsx'),
   require('../../vibrant-components/src/lib/Backdrop/Backdrop.stories.tsx'),
   require('../../vibrant-components/src/lib/Body/Body.stories.tsx'),
   require('../../vibrant-components/src/lib/BreadCrumb/BreadCrumb.stories.tsx'),
@@ -39,6 +40,8 @@ export const stories = [
   require('../../vibrant-components/src/lib/Paragraph/Paragraph.stories.tsx'),
   require('../../vibrant-components/src/lib/PasswordField/PasswordField.stories.tsx'),
   require('../../vibrant-components/src/lib/Pressable/Pressable.stories.tsx'),
+  require('../../vibrant-components/src/lib/Radio/Radio.stories.tsx'),
+  require('../../vibrant-components/src/lib/RadioGroupField/RadioGroupField.stories.tsx'),
   require('../../vibrant-components/src/lib/RangePickerField/RangePickerField.stories.tsx'),
   require('../../vibrant-components/src/lib/Ratio/Ratio.stories.tsx'),
   require('../../vibrant-components/src/lib/SafeAreaView/SafeAreaView.stories.tsx'),