+++ packages/vibrant-utils/src/index.ts
@@ -26,6 +26,5 @@ export { useControllableState } from './lib/useControllableState';
 export { useIsomorphicLayoutEffect } from './lib/useIsomorphicLayoutEffect';
 export { useObjectMemo } from './lib/useObjectMemo';
 export { useComposedRef } from './lib/useComposedRef';
-export { action } from './lib/storybook';
 export type { Position } from './types/Position';
 export type { Either } from './types/Either';