+++ packages/vibrant-components/CHANGELOG.md
@@ -2,6 +2,55 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.58.1](https://github.com/pedaling/opensource/compare/vibrant-components-0.58.0...vibrant-components-0.58.1) (2023-05-04)
+
+## [0.58.0](https://github.com/pedaling/opensource/compare/vibrant-components-0.57.1...vibrant-components-0.58.0) (2023-05-04)
+
+
+### Bug Fixes
+
+* useWindowDimensions훅을 최상위뷰의 width, height를 리턴하게 변경한다 ([#700](https://github.com/pedaling/opensource/issues/700)) ([ea7e39e](https://github.com/pedaling/opensource/commit/ea7e39e5e942a1708949ebc23be6bdbd7992ccd4))
+
+### [0.57.1](https://github.com/pedaling/opensource/compare/vibrant-components-0.57.0...vibrant-components-0.57.1) (2023-05-03)
+
+
+### Performance Improvements
+
+* Motion의 사용성을 개선한다 ([#699](https://github.com/pedaling/opensource/issues/699)) ([a25a734](https://github.com/pedaling/opensource/commit/a25a7344c1fa5d682e6f6d7031584d90bc5d3ed2))
+
+## [0.57.0](https://github.com/pedaling/opensource/compare/vibrant-components-0.56.0...vibrant-components-0.57.0) (2023-05-02)
+
+
+### Features
+
+* Motion 컴포넌트를 motion 라이브러리를 사용해서 구현한다  ([#692](https://github.com/pedaling/opensource/issues/692)) ([b997bbc](https://github.com/pedaling/opensource/commit/b997bbcf9bb314b02b90e4de93da98e619b819a9))
+
+## [0.56.0](https://github.com/pedaling/opensource/compare/vibrant-components-0.55.0...vibrant-components-0.56.0) (2023-04-26)
+
+## [0.55.0](https://github.com/pedaling/opensource/compare/vibrant-components-0.54.2...vibrant-components-0.55.0) (2023-04-25)
+
+
+### Features
+
+* Radio, RadioGroupField 컴포넌트를 추가한다 ([#683](https://github.com/pedaling/opensource/issues/683)) ([07a77fa](https://github.com/pedaling/opensource/commit/07a77fad4fbe3d49c5629088ccfd5ddcc81a7ca8))
+
+
+### Bug Fixes
+
+* add tab-index prop to pressable ([#684](https://github.com/pedaling/opensource/issues/684)) ([7c8a0be](https://github.com/pedaling/opensource/commit/7c8a0bedff10933c0f2d5e9025911aa2933138a6))
+* callout icon and contents margin ([#689](https://github.com/pedaling/opensource/issues/689)) ([2bf9ab4](https://github.com/pedaling/opensource/commit/2bf9ab4230b00cca95459567ee9ba63b580fae22))
+* change the operator used by TableMultiSelectFilter ([#682](https://github.com/pedaling/opensource/issues/682)) ([7d26611](https://github.com/pedaling/opensource/commit/7d26611c5e4ec9390e06ad7458ace801a2741fbc))
+* change unchecked radio icon button to outline1 ([#685](https://github.com/pedaling/opensource/issues/685)) ([5f1cca5](https://github.com/pedaling/opensource/commit/5f1cca541133f23b3e0d88d309da0e382e35fd26))
+
+### [0.54.2](https://github.com/pedaling/opensource/compare/vibrant-components-0.54.1...vibrant-components-0.54.2) (2023-04-17)
+
+
+### Bug Fixes
+
+* remove storybook from utils package ([#680](https://github.com/pedaling/opensource/issues/680)) ([c569967](https://github.com/pedaling/opensource/commit/c569967595afc6d2eabde98fbe32f9db864c19ab))
+
+### [0.54.1](https://github.com/pedaling/opensource/compare/vibrant-components-0.54.0...vibrant-components-0.54.1) (2023-04-13)
+
 ## [0.54.0](https://github.com/pedaling/opensource/compare/vibrant-components-0.53.7...vibrant-components-0.54.0) (2023-04-12)