+++ packages/workspace-plugin/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.40.1](https://github.com/pedaling/opensource/compare/workspace-plugin-0.40.0...workspace-plugin-0.40.1) (2023-05-04)
+
+## [0.40.0](https://github.com/pedaling/opensource/compare/workspace-plugin-0.39.1...workspace-plugin-0.40.0) (2023-05-04)
+
+### [0.39.1](https://github.com/pedaling/opensource/compare/workspace-plugin-0.39.0...workspace-plugin-0.39.1) (2023-05-03)
+
+## [0.39.0](https://github.com/pedaling/opensource/compare/workspace-plugin-0.38.0...workspace-plugin-0.39.0) (2023-05-02)
+
+## [0.38.0](https://github.com/pedaling/opensource/compare/workspace-plugin-0.37.0...workspace-plugin-0.38.0) (2023-04-26)
+
+## [0.37.0](https://github.com/pedaling/opensource/compare/workspace-plugin-0.36.2...workspace-plugin-0.37.0) (2023-04-25)
+
+### [0.36.2](https://github.com/pedaling/opensource/compare/workspace-plugin-0.36.1...workspace-plugin-0.36.2) (2023-04-17)
+
+### [0.36.1](https://github.com/pedaling/opensource/compare/workspace-plugin-0.36.0...workspace-plugin-0.36.1) (2023-04-13)
+
 ## [0.36.0](https://github.com/pedaling/opensource/compare/workspace-plugin-0.35.7...workspace-plugin-0.36.0) (2023-04-12)
 
 ### [0.35.7](https://github.com/pedaling/opensource/compare/workspace-plugin-0.35.6...workspace-plugin-0.35.7) (2023-04-11)